Move nova tests to `proof/nova_tests.rs`
