+++ src/proof/mod.rs
@@ -9,6 +9,8 @@
 pub mod groth16;
 /// An adapter to a Nova proving system implementation.
 pub mod nova;
+/// Nova tests
+pub mod nova_tests;
 
 /// An adapter to a SuperNova proving system implementation.
 pub mod supernova;
